Ported to use `cqrs_tools`
Changes

* Commands

  All commands now use Cqrs.Command and Cqrs.Command.EventDerivation

  `Cqrs.Command` provides a typed struct (Ecto embedded_schema) with the
  following functions.

  - `new` is constructor function that will validate incoming data.

    You can pass some options to `new`. A nice one is `return:
    :context`. With that you can view the whole history of the dispatch
    and some other interesting data.

  - `dispatch` will run the default Cqrs.DispatchStrategy. We've
  customized this to simply dispatch all commands to Commanded. see
  `Honeydew.Cqrs.PipelineResolver`

  `Cqrs.Command.EventDerivation` provides a `derive_event` macro to
  create structs based off of the current command's fields.

  You can also add more fields in the body.
  You can also pass the `drop: [:field1, :field2]` option to drop those
  fields from the resulting event.

* Queries

All queries are now using `Cqrs.Query` and define a seperate
`Cqrs.QueryPipeline` module.

The Query module defines any fields and bindings supported by
the query.

`fields` and the same as commands
`bindings` are for documenting to the caller that a query supports these
bindings. This is usefull when you don't want to execute the query, but
allow it to be used in a larger composed query.

Queries are generated with the functions as a Command.

If you send `execute: false` to a query's `dispatch`, only the query
will be returned.

* Context file

`cqrs_tools_ddd` supplies us with the `Cqrs.BoundedContext` module.

Using `Cqrs.BoundedContext` provides to macros.

- `command CommandModule`
- `query QueryModule`

These will generate proxy functions to dispatch Commands and queries and
to create a query.